Add title page for the University of Tübingen
The mathematical scientific faculty of the University of Tübingen makes clear at
this location

http://www.uni-tuebingen.de/index.php?eID=tx_nawsecuredl&u=0&g=0&t=1474369729&hash=2af453bd904baafee31f4765d710626a40bbec98&file=fileadmin/Uni_Tuebingen/Fakultaeten/MatNat/Dekanat/Promotion/Titelblatt_Muster.pdf

how the format of the title page should be.  This packages address problem.  It
loads babel.sty with options english and ngerman.  This package has been tested
for DIN A5 paper.  For DIN A4 there should be a different work around.